Fix for extract issue with leading/trailing spaces in directory names.
When extracting directories with names containing spaces, the files in
these directores could not be extracted.

The fix is to check that the created directory name is really equal to
the name that we intended to create (method deepEquals), and to fail if
it isn't, requiring the user to rename the directory to something sane.