Java6Specific.java: Use reflection to find Java 6-specific methods.
These methods can't be implemented as stubs since they are part of
already existing classes (java.awt.Window and javax.swing.JTable). We
can't 'extend' base classes like that, so the only way to get this to
compile on Java 5 is to switch to using reflection to find and invoke
the method.

Since this is expected to succeed, we throw a RuntimeException for each
exception type thrown in the process.