CommonHFSExtentKey.java: Fixed incorrect key compare order.
The specification states that the key fields should be compared in the
following order:
  1. File ID.
  2. Fork type.
  3. Start block.

However we compared the key fields in the order they appear in the
struct:
  1. Fork type.
  2. File ID.
  3. Start block.

Since resource forks very rarely have overflowing extents, we haven't
run into this problem yet in practice (fork type would always be the
same, i.e. the data fork), but this is clearly a bug that could prevent
us from properly descending through the extents overflow B-tree.

Note that this was a bug specific to the 'HFS common' layer. The HFS+
specific key compare function in HFSPlusExtentKey that we used before
the HFS+/HFS unification did the comparison in the proper order.