ExtentsInfoPanel.java: Changed display order of key fields in titles.
We recently fixed the comparison algorithm for overflow extents file
keys so it compares the fields in the correct order. This order should
also be reflected in the display order of the keys in the node titles in
ExtentsInfoPanel, which should be an accurate view of the keys' natural
order.