Add new utility, 'DumpHfs'.
This dumps all the critical file system metadata for the volume to
standard output, but leaves all data out (data blocks are written zeroed
to standard output).

Knows issues: It currently does not include the contents of symlinks,
and no extended attributes. It's debatable whether these should be
included, but at least symlink support should be added in the future
since this affects the file structure from a UNIX standpoint.