DumpHfs.java: Fix broken assumption about the end of the volume.
We assumed that the end of the volume was always 512 bytes after the end
of the last allocation block. However, the end of the volume could be up
to one allocation block minus one sector after the end of the last
allocation block.

Sadly HFS+ does not specify where the volume ends in the volume header,
so we must determine it ourselves, by reading block by block at the end
of the volume until we can't read anymore.