CommonHFSAttributesKey.java: Fix key comparison algorithm.
The key comparison algorithm was flawed and always compared the length
of the attribute names before comparing the attribute name data. The
proper way is to compare the attribute names up to the smallest of the
two name lengths. If the names still match up to that point, then
compare the attribute name lengths and let the shorter name sort before
the longer.