InternalStructViewPanel.java: Omit border on last element.
The border is there to separate this element from the next element in
the layout. If there is no next element, then the border will just be
useless empty space after the last element. This doesn't look good,
especially in the case of nested struct view panels, where the empty
space will grow with 5 pixel for each nested level.