VolumeInfoWindow.java: Deprecate usage of HFSPlusVolumeInfoPanel.
HFSPlusVolumeInfoPanel is a manually constructed view of the fields in
the volume header and was constructed before the advent of
StructElements, StructViewPanel and friends.

Now StructViewPanel has matured to the point where it can completely
replace HFSPlusVolumeInfoPanel, and in fact displays more information.
Is also now looks pretty much identical to the old HFSPlusVolumePanel
(in fact it looks better, with indentation being more consistent).