Update csframework.jar with recent bugfix.
The following bugfix was imported from the framework:
========================================================================
Fix bug in 'readFrom' where file pointer was left changed on exception.

We should always try to seek back to the old file pointer, even when
there's an exception. Anything else would be a violation of the
semantics of SynchronizedReadableRandomAccessStream. So surround the
'read' statement with 'try {}' and put the final seek in a 'finally {}'
block.
========================================================================