ReadableBlockCachingStream.java: Allow seeking to the end of the file.
This is ok, but a subsequent read of non-0 length will fail. If we have
read all the bytes in the stream, we will end up with a file pointer
that points to the byte after the end of the file, and even in this
situation we must be able to do:
    <stream>.seek(<stream>.getFilePointer());
This is something that did not work previously.