Implement Java 7-specific method of setting more file times on extract.
When extracting files in HFSExplorer, we now set both the last accessed,
last modified and created timestamps for the file (where this is
supported by the OS).

This is only possible in Java 7, and since we want the software to work
on earlier Java versions we use reflection in order to access this
functionality.