Java6Specific.java: Eliminate lazy classloading in Java6Specific.
In order to be able to compile the software on Java 5, even if we have
Java 6-specific code, we relied on lazy class loading and stub class
files when compiling to fill in the gaps. This is really not very
elegant, so we change the lazy class loading method to instead use
reflection to invoke the necessary Java 6 functionality. This way the
code can be compiled cleanly in Java 5 without relying on Java 6 stub
classes.