Update csframework.jar with recent bugfix for Java 6 detection.
The following bugfix was imported from the framework:
========================================================================
Java6Util.java: Fix incorrect property checked when checking for Java 6.

The property 'java.vm.version' is really irrelevant to whether the VM
supports Java 6, being an internal Java VM version number.
What we are really after is the version of the Java specification that
is supported by the VM, and this is stored in
'java.specification.version'.
========================================================================