MacSpecific.java: Eliminate lazy class loading in Mac-specific code.
In order to be able to compile the software on non-Mac OS X platforms,
we relied on lazy class loading and stub class files when compiling to
fill in the gaps. This is really not very elegant, so we change the lazy
class loading method to instead use reflection to invoke the necessary
Mac OS X-specific functionality. This way the code can be compiled
cleanly on non-Mac OS X platforms without relying on stub classes for
Apple's Java extensions.