FileSystemBrowserWindow.java: Emit extract error message to System.err.
The error message is returned to the caller and displayed when
extraction is complete. However in order to match exceptions printed to
System.err with error messages displayed at the end of the extract
process, we should also emit the same message to System.err. Otherwise
it's difficult to find the corresponding exception stack trace.