Split setting of Java 7 POSIX mode and ownership into separate stages.
Changing mode of symbolic links cannot be done for some reason, while
changing ownership works fine. This meant that when attempting to do
everything at once, we got an exception when attempting to set the mode,
which then blocked the setting of ownership. Solved by splitting the two
operations into separate functions.