FileSystemBrowserWindow.java: Set attributes for directory on creation.
This is done to permit entries whose attributes for some reason could
not be modified to assume some sane default values taken from its parent
directory.

Also this means that the owner of the files can browse them while they
are being extracted (in case another user is extracting them, e.g. root
if this is required in order to access the device).