Add support for AES-256 encrypted disk images to HFSExplorer.
In order to achieve this functionality, hfsx_dmglib.jar was updated to
include the following commit (and its ancestors):
========================================================================
Add support for AES-256 encrypted disk images to DMGExtractor.

CommonCEncryptedEncodingHeader.V2Implementation.unwrapKeys(...) was
extended to account for variable key sizes and the unknown variable at
offset 24 in the V2Header header was found to store the key size in
bits.

We earlier made V2Header a variable-sized struct, so that the key is
properly stored even when it's larger than 16 bytes. Together these
modifications are sufficient to enable working support for AES-256
encrypted disk images.
========================================================================

Also csframework.jar was updated as a dependency to include the
following commit (and its ancestors):
========================================================================
Add new method IOUtil.readFully(ReadableRandomAccessStream, long, int).
========================================================================