Fix leakage of open streams in HFSExplorer.
Open streams were leaked freely in HFSExplorer, causing problems when
closing a volume and then opening it again since the volume would still
be open but leaked pending garbage collection (the file locks would
still be active).

To fix this we update the framework to include the change to "lazy
close" for SychronizedReadableRandomAccessStream and implement missing
close() calls in various layers in HFSExplorer.
We also reengineer the stream handling to rely on lazy closing and
making sure references are kept active when needed at all time.

The framework was updated to include the following commit and its
ancestors:
========================================================================
SynchronizedReadableRandomAccessStream.java: Add warning in finalize().

If reference count on invocation of finalize() is non-0, we emit a
warning to alert that all the substreams have not been properly closed,
i.e. we are leaking the source stream in an open state.
========================================================================