Correct variable names / struct size of DriverDescriptorRecord.
The variable names for the struct members that we call "reserved1",
"reserved2" and "reserved3" are "sbDevType", "sbDevId" and "sbData" in
Apple's documentation.

In addition we fix the formatting of the "Description" messages, change
the struct size in the documentation to reflect that the struct size is
not fixed and remove the mention of 30 driver descriptor entries in the
struct description (the actual number of entries are described by
"sbDrvrCount").