launcher.cpp: Add code to open Java download page on launch error.
The dialog presented to the user used to contain only one button "OK",
and this button simply closed the dialog and exited the launcher.

We now have two buttons, "OK" and "Cancel". When clicking "OK" the user
is taken to http://www.java.com using the default web browser ("Cancel"
exits the launcher like before).

The message displayed to the user is also updated accordingly, and all
references to the old download site http://java.sun.com (which nowadays
redirects to Oracle's Java download page) is removed.

Also added some icons to the dialogs.