Retire the old BuildEnumerator system.
It has not been used at all lately, and frankly it should not be used.
It has been causing problems with build number being updated downstream
causing the version check to fail. In the future we will update the
build number manually, incrementing it with some fixed number on each
release.