Fix issue with UDIF files having non-ASCII characters in their XML data.
hfsx_dmglib.jar was updated to include the following commit (and its
ancestors):
========================================================================
CharByCharReader.java: Fix incorrect buffer size for decoding.

If for example UTF-8 is used, then non-ASCII characters would lead to an
exception when decoding because of incorrect length of input buffer.

The buffer length should be based on maximum bytes per char, not maximum
chars per byte. Big difference. The solution is to create a temporary
encoder for the charset when creating the CharByCharReader just to get
the maxBytesPerChar() value, and use this value for the array length of
tempBuffer.
========================================================================