Fix problem when searching attribute leaf nodes with findLEKeys.
When searching leaf nodes with findLEKeys, the record before the first
matching record was always returned. This is correct when searching
index nodes because the nodes further down the hierarchy might contain
the nodes that we are searching for. When searching leaf nodes though we
always want only the results strictly matching the search conditions.