Fix sparse bundle locking issues with the Boot Camp Windows HFS+ driver.
hfsx_dmglib.jar was updated to include the following commit (and its
ancestors):
========================================================================
SparseBundle.java: Downgrade locking errors to only trigger a warning.

Apple's Boot Camp Windows HFS+ file system driver doesn't support
locking, so an exception is thrown when attempting to load sparse
bundles from an HFS+ volume in Windows (Boot Camp).

It seems that the only solution is to consider locking of the files in
the sparse bundle optional. This increases the possibility of concurrent
modification, but really this should be a non-issue in most cases
anyway.
========================================================================