Move POSIX filename translation into the file system handler.
A new handler factory custom attribute, "POSIX_FILENAMES", was created
to control whether or not POSIX filename translation (':' <-> '/')
should be applied to path components and UnHFS was updated as the first
utility to take advantage of this functionality.

The attribute is opt-in, i.e. the default is still to leave this POSIX
translation turned off and return file names in the exact same form as
they are stored.

This also meant that I had to fix broken/non-existent Unicode
decomposition in order to get a proper round-trip conversion. Filenames
used to be returned from the handler in composed form but during lookup
there was no decomposition.