HFSCommonFSEntry.java: Ignore some fields in empty Finder info check.
For consistency with XNU's kernel file system driver, we zero certain
fields before checking whether the Finder info data is to be considered
empty.

If considered empty, the Finder info data will not be accessible through
the com.apple.FinderInfo attribute (until it is set to something
considered to be non-empty).