Indicate compressed files with blue colour in HFSExplorer.
We substitute the colour black with blue when painting the filename
component in the JTable. Unconditionally setting the whole thing to blue
led to unexplained painting problems so we only do it when the original
colour is black. This may break this functionality in some UI L&Fs where
the default text colour is not black, but it has to be that way until I
figure out the painting issues.