Fix partially broken symlink functionality in HFSExplorer.
During the refactoring that was done when adding attributes file
functionality, symlink functionality in HFSExplorer was inadvertently
broken.

The problem was that the HFSCommonFSLink class was refactored to derive
from HFSCommonFSFile in order to share common code... however this meant
that it was both an instance of FSLink and FSFile, which should not be
the case and which breaks certain code testing whether an FSEntry is a
file or link.

This was fixed by creating an abstract superclass of both
HFSCommonFSFile and HFSCommonFSLink, which implements only FSEntry.