Add HFSCommonFSFile.java, missing from last commit.
This is intentional, since git isn't smart enough to figure out when you
have moved a file and when a file is new.

Having this file in the previous commit breaks git blame functionality
for HFSCommonAbstractFile.java since git thinks is was created starting
with the previous commit. Hence this hack.