Add new applesingle structs AttributeEntry / AttributeHeader.
These structs are related to storage of extended attributes inside an
applesingle format file (in practice this probably only applies to
AppleDouble files).

Extended attributes are stored inside the Finder info section, right
after the Finder info data (16 bytes of Finder info + 16 bytes of
extended Finder info + 2 bytes of padding).
It stars with an attribute header (AttributeHeader.struct), followed by
attribute entry descriptors (AttributeEntry.struct) containing the name
of the attribute and the offset/length of the attribute data relative to
the start of the file.