Check compressed flag to determine whether a file is compressed.
This should be the first test, before we start probing the attributes
file. If the compressed flag in the BSD flags is not set, then the file
is definitely not compressed.

This change should improve performance since we don't have to peek into
the attribute file to check for the decmpfs attribute every time we look
up a file.