FileSystemBrowserWindow.java: Fix bad deepExists behaviour for root dir.
The function deepExists tries to verify the existence of a file by
listing its parent directory and matching its name against the entries
in its parent. However when there is no parent, which is the case for
the root directory, a NullPointerException would be thrown.

Fixed by returning true if no parent exists.