HFSCommonFileSystemRecognizer.java: Don't throw exception on empty file.
If a file (device) was smaller than 4096 bytes, an exception was thrown
by the HFS common file system recognizer since readFully didn't manage
to fully read the requested number of bytes.

Fixed by using the regular 'read' method and returning normally with
FileSystemType.UNKNOWN if the requested number of bytes could not be
read.