Fix issues with parsing the encryption data in the version 2 header.
The structure of the version 2 header was poorly understood, and fields
assumed to be static were in fact dynamically laid out. This fix should
vastly improve compatibility with images encrypted with the version 2
CEncryptedEncoding scheme.

This updates hfsx_dmglib.jar to include all commits up to and including:
========================================================================
V2Header.java: Emit warnings for unexpected values of fields.

Some field members in the V2 header are not yet fully understood. Since
we don't know the exact meaning of these fields we now emit warning
messages on System.err when they deviate from what we have seen
previously.
This helps the developer to get valuable feedback for non-standard
images, and hopefully to understand the meaning of the fields.
========================================================================