Fix incorrect unsigned integer max limit checks in the HFS library.
Some limit checks were performed by comparing against the limit of the
signed type multiplied with 2. However the result of this calculation is
1 less than the actual limit of the unsigned type (e.g. Short.MAX_VALUE
is 32767, double this and you get 65534, not 65535 which is the actual
limit of a 16-bit unsigned integer).

Fixed by gathering all limits in a new 'Limits' interface in the
org.catacombae.hfs package (with constant names matching the ones in the
C99 standard) and using them.