Improve error handling when processing HFS+ catalog leaf node structs.
A new HFSException base class is introduced, intended to be used only
with exceptions that arise due to invalid on-disk data. Since we can now
distinguish between those exceptions and other errors (programming
errors, errors from the I/O layer) we can also handle them more
gracefully and instead of throwing an exception all the way to the
initial caller we can handle the issue where this is possible.

The main difference from before is that individual records in a leaf
node can now be corrupted without halting the entire process of
processing the node.