Fix out of bounds read bug in concatenated stream.
csframework.jar is updated to include the following commit (and its
ancestors):
========================================================================
ConcatenatedStream.java: Fix writing beyond bounds of stream.

The logic in BasicConcatenatedStream.write(byte[], int, int) to skip to
the part containing the first byte of the write was flawed if the file
pointer pointed at the final byte of the concatenated stream, in which
case we could just keep writing beyond its bounds, advancing the file
pointer far beyond the defined length of the stream.

Fixed by properly detecting when we are writing beyond the last part and
returning -1 if we do.
========================================================================