Add ARM/ARM64 builds of llio.
This is requrired for HFSExplorer to work with actual devices on
ARM64-native Java, like the early versions published by Microsoft at:
  https://github.com/microsoft/openjdk-aarch64

As for 32-bit ARM, I don't think there ever was a Windows Java release
for this architecture but it's included for completeness in case I'm
wrong.

Building for ARM targets currently requires Visual Studio 2019
(Community Edition is fine) as this is what I had available.