Add support for the MacJapanese encoding when parsing HFS volumes.
Switching between charsets is now implemented as a special combo box in
the UI which only shows up for HFS volumes. Changing a different charset
causes the UI to reload and lose the current context, but that's good
enough for now.

The MacJapanese encoding isn't commonly available in the charsets that
comes with Java or the system, so it is implemented as a custom
StringCodec using data from unicode.org.

Due to the size of the data and Java's size restrictions it had to be
split into three class files... it could probably be done in a nicer way
but this way it's compiled in and always available.