Dispose of the debug console window when the main window is closed.
This is the simplest solution to the debug console window blocking exit
of the application because it's not in a disposed state.

A slightly better solution would be to add a reference to it for every
window created by this HFSExplorer instance and dispose of it only when
the last window is closed but that would require a lot more work and
this is good enough for now.