Fix issues with single-byte codepage encoding.
A lot of the code was copied from MacJapaneseStringEncoding which is
more complex and has different requirements. For instance we don't need
to use a full String as key in the hash map, a Character is sufficent,
and there's no reason to check for 5 to 2-character substrings when
looking up the replacement as the mapping is one-to-one.