Fix incorrect partition data returned for non 512-byte block sizes.
APMPartition assumed a struct size of 512, but it varies with the block
size as specified in the driver descriptor record (a partition always
occupies a full block).
Thus the data returned by ApplePartitionMap.getData(...) was incorrect
and omitted the padding required to fill a whole block, instead laying
out the entries as if they were using 512 bytes as block size.
Fixed by making the pmPad field variable-sized and not static and
copying all the data of the partition entry.