AppleSingleBuilder.java: Fix incorrect 'totalEnd' in AttributeHeader.
This should be referring to the end of this entry and the start of the
next one, but with padding which is usually applied to the FinderInfo /
extended attributes entry, this isn't always accurate.
Since we didn't know how many bytes have been allocated with padding, we
add another 'length' parameter to the getBytes(...) prototype and pass
the allocated size down to the FinderInfoEntry logic.